% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extent.R
\name{extent}
\alias{extent}
\title{Extent is a workhorse, from this is derived
xlim,ylim,xmin,xmax,ymin,ymax and from those
x_res,y_res,x,y_centre,corner, etc.}
\usage{
extent(x)
}
\description{
Extent is a workhorse, from this is derived
xlim,ylim,xmin,xmax,ymin,ymax and from those
x_res,y_res,x,y_centre,corner, etc.
}
