% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granule.R
\name{granule}
\alias{granule}
\alias{granule.default}
\alias{granule.Extent}
\alias{granule.bbox}
\alias{granule.BasicRaster}
\alias{granule.stars}
\alias{granule.SpatRaster}
\title{Title}
\usage{
granule(
  xmin = 0,
  xmax = 1,
  ymin = 0,
  ymax = 1,
  ncol = 1,
  nrow = 1,
  crs = list()
)

\method{granule}{default}(
  xmin = 0,
  xmax = 1,
  ymin = 0,
  ymax = 1,
  ncol = 1,
  nrow = 1,
  crs = list()
)

\method{granule}{Extent}(
  xmin = 0,
  xmax = 1,
  ymin = 0,
  ymax = 1,
  ncol = 1,
  nrow = 1,
  crs = list()
)

\method{granule}{bbox}(
  xmin = 0,
  xmax = 1,
  ymin = 0,
  ymax = 1,
  ncol = 1,
  nrow = 1,
  crs = list()
)

\method{granule}{BasicRaster}(
  xmin = 0,
  xmax = 1,
  ymin = 0,
  ymax = 1,
  ncol = 1,
  nrow = 1,
  crs = list()
)

\method{granule}{stars}(
  xmin = 0,
  xmax = 1,
  ymin = 0,
  ymax = 1,
  ncol = 1,
  nrow = 1,
  crs = list()
)

\method{granule}{SpatRaster}(
  xmin = 0,
  xmax = 1,
  ymin = 0,
  ymax = 1,
  ncol = 1,
  nrow = 1,
  crs = list()
)
}
\arguments{
\item{crs}{}
}
\value{

}
\description{
Title
}
